MGMIDN-1046: Fix bug Leaderboard
Merge in DA-NANG-INTERNSHIP/easy-tourney-be from feature/MGMIDN-1046-sp-2-show-leaderboard-be to master

Squashed commit of the following:

commit 8aad3ba4c1529757473da5c4a1a895ac30f7835a
Author: Long Bien <lbit@pop-os.localdomain>
Date:   Thu Dec 28 09:35:06 2023 +0700

    MGMIDN-1046: Fix bug Leaderboard

+++ src/main/java/com/example/easytourneybe/match/MatchOfLeaderBoardDto.java
@@ -31,5 +31,5 @@ public class MatchOfLeaderBoardDto {
 
     private LocalTime endTime;
 
-    private Integer teamWinId;
+    private Long teamWinId;
 }
